requirements.txt
# 双极构念聚类分析系统依赖包
# Bipolar Constructs Clustering Analysis - Dependencies

# 核心计算包
numpy>=2.1.0
pandas>=2.2.0
scipy>=1.14.0

# 机器学习
scikit-learn>=1.6.0
umap-learn>=0.5.0

# 可视化
plotly>=6.0.0

# 可选：系统监控 (推荐安装)
# psutil>=5.8.0

# 可选：Jupyter notebook支持
# jupyter>=1.0.0
# ipywidgets>=7.6.0

# 安装命令:
# pip install -r requirements.txt
